Attract demo matches stable across seeds: bitfield unions, overlays, raw offsets, disc tables
- Runtime bitfield/integer unions marked DISC_STRUCT (union + nested structs)
  so bit order matches the GameCube (Fighter x594, MotionState, Struct2070,
  stage flag words, Big Blue lanes...)
- Padded overlay structs over runtime data replaced by real field access
  (izumi, ifstock, mnname user data, toy, gm results/tournament, kirby...)
- Remaining adjacent-static walks (ftmaterial templates, particle list,
  mndiagram, results screen, gm_180A, toy) replaced by direct references
- Raw GC byte offsets into Ground/menus (Big Blue, gmtoulib, toy) -> fields
- Disc data read natively fixed: item throw attrs, ftDynamics blend tables,
  Yoshi part tables, ItCollDynamics, zebes anim tables; DP() zero-extends
- GC-sized allocs/memzeros (ifmagnify, staffroll, mnname, mpisland, ...)
- tools/demo_sweep.sh; MELEE_HEAP_CHECK canaries in aurora OSAlloc

+++ src/pc/disc.h
@@ -42,7 +42,7 @@
 
 #define DISC_STRUCT __attribute__((scalar_storage_order("big-endian")))
 #define DISC_PTR(T) uint32_t
-#define DP(T, slot) ((T*) (uintptr_t) (slot))
+#define DP(T, slot) ((T*) (uintptr_t) (uint32_t) (slot)) /* zero-extend even if the slot is signed */
 
 void pc_disc_ptr_overflow(const void* p, const char* file, int line) __attribute__((noreturn));
 #define DP_SET(slot, p)                                                        \
